+++ lawaf/__init__.py
@@ -1,4 +1,16 @@
 from lawaf.interfaces import W90Downfolder, SiestaDownfolder, PhonopyDownfolder, NACPhonopyDownfolder
+from lawaf.interfaces.phonopy.symmetry_seeds import (
+    get_symmetry_anchor_wfn,
+    list_opd_families,
+)
 from lawaf.lwf.lwf import LWF
 
-__all__ = [W90Downfolder, SiestaDownfolder, PhonopyDownfolder, NACPhonopyDownfolder, LWF]
+__all__ = [
+    "W90Downfolder",
+    "SiestaDownfolder",
+    "PhonopyDownfolder",
+    "NACPhonopyDownfolder",
+    "LWF",
+    "get_symmetry_anchor_wfn",
+    "list_opd_families",
+]